requirements.txt
click
openai
PyGithub
pytest
python-dotenv 